requirements.txt
adal==1.2.2
applicationinsights==0.11.9
azure-common==1.1.25
azure-core==1.3.0
azure-graphrbac==0.61.1
azure-identity==1.3.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==8.0.1
azure-mgmt-storage==8.0.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.3.0
azureml-automl-core==1.2.0
azureml-core==1.2.0
azureml-dataprep==1.3.5
azureml-dataprep-native==14.1.0
azureml-pipeline==1.2.0
azureml-pipeline-core==1.2.0
azureml-pipeline-steps==1.2.0
azureml-sdk==1.2.0
azureml-telemetry==1.2.0
azureml-train==1.2.0
azureml-train-automl-client==1.2.0
azureml-train-core==1.2.0
azureml-train-restclients-hyperdrive==1.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
catboost==0.22
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
contextlib2==0.6.0.post1
#cryptography==2.8
cycler==0.10.0
DateTime==4.3
distro==1.4.0
docker==4.2.0
dotnetcore2==2.1.13
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
fusepy==3.0.1
glibc==0.6.1
graphviz==0.13.2
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonpickle==1.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msal==1.1.0
msal-extensions==0.1.3
msrest==0.6.11
msrestazure==0.6.3
ndg-httpsclient==0.5.1
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pathspec==0.7.0
plotly==4.6.0
portalocker==1.6.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.15.89
scikit-learn==0.21.3
scipy==1.4.1
SecretStorage==3.1.2
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
xlrd==1.2.0
zope.interface==5.0.2
